TPS camera behavior implemented
