+++ Assets/_Project/Scripts/Core/Character/PlayerInput.cs
@@ -163,7 +163,7 @@ public partial class @PlayerInput: IInputActionCollection2, IDisposable
                     ""id"": ""c642d864-52ec-41e9-ab99-023c897554b2"",
                     ""path"": ""<Mouse>/delta"",
                     ""interactions"": """",
-                    ""processors"": ""InvertVector2(invertX=false),ScaleVector2(x=0.1,y=0.1)"",
+                    ""processors"": ""ScaleVector2(x=0.1,y=0.1)"",
                     ""groups"": ""Keyboad"",
                     ""action"": ""Look"",
                     ""isComposite"": false,
