+++ Assets/_Project/Scripts/Core/Character/PlayerInput.inputactions
@@ -141,7 +141,7 @@
                     "id": "c642d864-52ec-41e9-ab99-023c897554b2",
                     "path": "<Mouse>/delta",
                     "interactions": "",
-                    "processors": "InvertVector2(invertX=false),ScaleVector2(x=0.1,y=0.1)",
+                    "processors": "ScaleVector2(x=0.1,y=0.1)",
                     "groups": "Keyboad",
                     "action": "Look",
                     "isComposite": false,
